+++ docs/astro.config.mjs
@@ -14,7 +14,7 @@ export default defineConfig({
     resolve: {
       alias: {
         "package:": "/src",
-        "@components": "@sv/components",
+        "@components": "@atrium-ui/components",
       },
     },
     server: {