+++ docs/src/content/docs/components/toast/ToastDemo.tsx
@@ -1,5 +1,5 @@
 /* @jsxImportSource vue */
-import "@atrium-ui/elements/portal";
+import "@sv/elements/portal";
 import { Button } from "@components/src/vue/Button";
 import { Toasts, toast } from "@components/src/vue/Toasts";
 import { paragraph } from "txtgen";