+++ packages/elements/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @sv/elements
+# @atrium-ui/elements
 
 ## 2.1.0 (2025-01-30)